requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
Django==3.2.23
django-allauth==0.57.0
django-cpf==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.5
